Moves TabPanel to its own component and Tab styles into the same folder.
