+++ src/ui/common/src/components/Tabs/Tabs.styles.ts
