+++ src/ui/common/src/components/integrations/dialogs/athenaDialog.tsx
@@ -2,7 +2,7 @@ import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
 import React, { useEffect, useState } from 'react';
 
-import { Tab, Tabs } from '../../../components/primitives/Tabs.styles';
+import { Tab, Tabs } from '../../Tabs/Tabs.styles';
 import {
   AthenaConfig,
   AWSCredentialType,
@@ -46,9 +46,9 @@ export const AthenaDialog: React.FC<Props> = ({
   const fileData =
     fileName && !!value?.config_file_content
       ? {
-          name: fileName,
-          data: value.config_file_content,
-        }
+        name: fileName,
+        data: value.config_file_content,
+      }
       : null;
 
   useEffect(() => {
